chore: remove dapp filter and search experiment
