+++ src/analytics/Properties.tsx
@@ -1142,8 +1142,7 @@ interface CICOEventsProperties {
 }
 
 interface DappProperties {
-  categoryId?: string
-  categories?: string[]
+  categories: string[]
   dappId: string
   dappName: string
 }
