+++ src/dapps/slice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { Actions as AppActions, UpdateConfigValuesAction } from 'src/app/actions'
-import { ActiveDapp, DappCategory, DappConnectInfo, DappV1, DappV2 } from 'src/dapps/types'
+import { ActiveDapp, Dapp, DappCategory, DappConnectInfo } from 'src/dapps/types'
 import { REMOTE_CONFIG_VALUES_DEFAULTS } from 'src/firebase/remoteConfigValuesDefaults'
 import { getRehydratePayload, REHYDRATE, RehydrateAction } from 'src/redux/persist-helper'
 
@@ -10,7 +10,7 @@ export interface State {
   maxNumRecentDapps: number
   recentDappIds: string[]
   dappListApiUrl: string | null
-  dappsList: Array<DappV1 | DappV2>
+  dappsList: Dapp[]
   dappsListLoading: boolean
   dappsListError: string | null
   dappsCategories: DappCategory[]
@@ -43,7 +43,7 @@ export interface DappSelectedAction {
 }
 
 export interface FetchDappsListCompletedAction {
-  dapps: Array<DappV1 | DappV2>
+  dapps: Dapp[]
   categories: DappCategory[]
   mostPopularDappIds: string[]
 }
