+++ src/dapps/types.ts
@@ -7,35 +7,22 @@ export enum DappSection {
   MostPopular = 'mostPopular',
 }
 
-export interface DappV1 {
+export interface Dapp {
   id: string
   iconUrl: string
   name: string
   description: string
   dappUrl: string
-  isFeatured: boolean
-  categoryId: string
-}
-
-export interface DappV2 {
-  id: string
-  iconUrl: string
-  name: string
-  description: string
-  dappUrl: string
-  isFeatured: boolean
   categories: string[]
 }
 
-export interface DappV2WithCategoryNames extends DappV2 {
+export interface DappV2WithCategoryNames extends Dapp {
   categoryNames: string[]
 }
 
-export const isDappV2 = (dapp: DappV1 | DappV2): dapp is DappV2 => 'categories' in dapp
-
 // Needs to be a type as an interface can only extend an object type
 // or intersection of object types with statically known members.
-export type ActiveDapp = (DappV1 | DappV2) & { openedFrom: DappSection }
+export type ActiveDapp = Dapp & { openedFrom: DappSection }
 
 export interface DappCategory {
   backgroundColor: string
