+++ src/dappsExplorer/DAppsBottomSheet.tsx
@@ -3,14 +3,14 @@ import { useTranslation } from 'react-i18next'
 import { StyleSheet, Text, View } from 'react-native'
 import BottomSheetLegacy from 'src/components/BottomSheetLegacy'
 import Button from 'src/components/Button'
-import { DappV1, DappV2 } from 'src/dapps/types'
+import { Dapp } from 'src/dapps/types'
 import QuitIcon from 'src/icons/QuitIcon'
 import { TopBarIconButton } from 'src/navigator/TopBarButton'
 import fontStyles from 'src/styles/fonts'
 import { Spacing } from 'src/styles/styles'
 
 interface Props {
-  selectedDapp: DappV1 | DappV2 | null
+  selectedDapp: Dapp | null
   isVisible: boolean
   onClose(): void
   onConfirmOpenDapp(): void
