+++ src/dappsExplorer/DAppsExplorerScreenSearchFilter.test.tsx
@@ -5,19 +5,19 @@ import { DappExplorerEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { dappSelected, favoriteDapp, fetchDappsList, unfavoriteDapp } from 'src/dapps/slice'
 import { DappCategory, DappSection } from 'src/dapps/types'
-import DAppsExplorerScreenFilter from 'src/dappsExplorer/filter/DAppsExplorerScreenFilter'
+import DAppsExplorerScreenSearchFilter from 'src/dappsExplorer/DAppsExplorerScreenSearchFilter'
 import { createMockStore } from 'test/utils'
-import { mockDappListV2 } from 'test/values'
+import { mockDappListWithCategoryNames } from 'test/values'
 
 jest.mock('src/analytics/ValoraAnalytics')
 jest.mock('src/statsig', () => ({
   getExperimentParams: () => ({
     dappsFilterEnabled: true,
-    dappsSearchEnabled: false,
+    dappsSearchEnabled: true,
   }),
 }))
 
-const dappsList = mockDappListV2
+const dappsList = mockDappListWithCategoryNames
 
 const dappsCategories: DappCategory[] = [
   {
@@ -38,7 +38,7 @@ const defaultStore = createMockStore({
   dapps: { dappListApiUrl: 'http://url.com', dappsList, dappsCategories },
 })
 
-describe(DAppsExplorerScreenFilter, () => {
+describe(DAppsExplorerScreenSearchFilter, () => {
   beforeEach(() => {
     defaultStore.clearActions()
     jest.clearAllMocks()
@@ -47,7 +47,7 @@ describe(DAppsExplorerScreenFilter, () => {
   it('renders correctly and fires the correct actions on press dapp', () => {
     const { getByText, queryByText } = render(
       <Provider store={defaultStore}>
-        <DAppsExplorerScreenFilter />
+        <DAppsExplorerScreenSearchFilter />
       </Provider>
     )
 
@@ -68,7 +68,7 @@ describe(DAppsExplorerScreenFilter, () => {
   it('renders correctly and fires the correct actions on press deep linked dapp', () => {
     const { getByText, queryByText } = render(
       <Provider store={defaultStore}>
-        <DAppsExplorerScreenFilter />
+        <DAppsExplorerScreenSearchFilter />
       </Provider>
     )
 
@@ -88,7 +88,7 @@ describe(DAppsExplorerScreenFilter, () => {
   it('displays the dapps disclaimer bottom sheet when selecting a dapp', () => {
     const { getByText } = render(
       <Provider store={defaultStore}>
-        <DAppsExplorerScreenFilter />
+        <DAppsExplorerScreenSearchFilter />
       </Provider>
     )
 
@@ -118,7 +118,7 @@ describe(DAppsExplorerScreenFilter, () => {
     })
     const { getByText, queryByText } = render(
       <Provider store={store}>
-        <DAppsExplorerScreenFilter />
+        <DAppsExplorerScreenSearchFilter />
       </Provider>
     )
 
@@ -146,7 +146,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -166,11 +166,11 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, queryByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
-      const favoritesSection = getByTestId('DAppsExplorerScreenFilter/FavoriteDappsSection')
+      const favoritesSection = getByTestId('DAppsExplorerScreenSearchFilter/FavoriteDappsSection')
       expect(within(favoritesSection).queryByText(dappsList[0].name)).toBeFalsy()
       expect(within(favoritesSection).getByText(dappsList[1].name)).toBeTruthy()
       expect(within(favoritesSection).getByText(dappsList[1].description)).toBeTruthy()
@@ -189,14 +189,14 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
       // don't include events dispatched on screen load
       jest.clearAllMocks()
 
-      const allDappsSection = getByTestId('DAppsExplorerScreenFilter/DappsList')
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
       fireEvent.press(within(allDappsSection).getByTestId('Dapp/Favorite/dapp2'))
 
       // favorited dapp confirmation toast
@@ -225,7 +225,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, getAllByTestId, queryByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -236,7 +236,7 @@ describe(DAppsExplorerScreenFilter, () => {
       // should only appear once, in the favorites section
       expect(selectedDappCards).toHaveLength(1)
 
-      const favoritesSection = getByTestId('DAppsExplorerScreenFilter/FavoriteDappsSection')
+      const favoritesSection = getByTestId('DAppsExplorerScreenSearchFilter/FavoriteDappsSection')
       fireEvent.press(within(favoritesSection).getByTestId('Dapp/Favorite/dapp2'))
 
       expect(queryByText('dappsScreen.favoritedDappToast.message')).toBeFalsy()
@@ -249,6 +249,103 @@ describe(DAppsExplorerScreenFilter, () => {
         section: 'all',
       })
       expect(store.getActions()).toEqual([fetchDappsList(), unfavoriteDapp({ dappId: 'dapp2' })])
+    })
+  })
+
+  describe('searching dapps', () => {
+    it('renders correctly when there are no search results', () => {
+      const store = createMockStore({
+        dapps: {
+          dappListApiUrl: 'http://url.com',
+          dappsList,
+          dappsCategories,
+          dappFavoritesEnabled: true,
+          favoriteDappIds: [],
+        },
+      })
+
+      const { getByTestId, queryByTestId } = render(
+        <Provider store={store}>
+          <DAppsExplorerScreenSearchFilter />
+        </Provider>
+      )
+
+      fireEvent.changeText(getByTestId('SearchInput'), 'iDoNotExist')
+
+      // Should display just the no results within the favorites section
+      expect(getByTestId('FavoriteDappsSection/NoResults')).toBeTruthy()
+      expect(queryByTestId('DAppsExplorerScreenSearchFilter/NoResults')).toBeNull()
+    })
+
+    it('renders correctly when there are search results in both sections', () => {
+      const store = createMockStore({
+        dapps: {
+          dappListApiUrl: 'http://url.com',
+          dappsList,
+          dappsCategories,
+          dappFavoritesEnabled: true,
+          favoriteDappIds: ['dapp2'],
+        },
+      })
+
+      const { getByTestId, queryByTestId } = render(
+        <Provider store={store}>
+          <DAppsExplorerScreenSearchFilter />
+        </Provider>
+      )
+
+      fireEvent.changeText(getByTestId('SearchInput'), 'dapp')
+
+      // Should display the correct sections
+      const favoritesSection = getByTestId('DAppsExplorerScreenSearchFilter/FavoriteDappsSection')
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
+
+      // Names display correctly in the favorites section
+      expect(within(favoritesSection).queryByText(dappsList[0].name)).toBeFalsy()
+      expect(within(favoritesSection).getByText(dappsList[1].name)).toBeTruthy()
+
+      // Names display correctly in the all dapps section
+      expect(within(allDappsSection).getByText(dappsList[0].name)).toBeTruthy()
+
+      // No results sections should not be displayed
+      expect(queryByTestId('FavoriteDappsSection/NoResults')).toBeNull()
+      expect(queryByTestId('DAppsExplorerScreenSearchFilter/NoResults')).toBeNull()
+    })
+
+    it('clearing search input should show all dapps', () => {
+      const store = createMockStore({
+        dapps: {
+          dappListApiUrl: 'http://url.com',
+          dappsList,
+          dappsCategories,
+          dappFavoritesEnabled: true,
+          favoriteDappIds: ['dapp2'],
+        },
+      })
+
+      const { getByTestId } = render(
+        <Provider store={store}>
+          <DAppsExplorerScreenSearchFilter />
+        </Provider>
+      )
+
+      // Type in search that should have no results
+      fireEvent.press(getByTestId('SearchInput'))
+      fireEvent.changeText(getByTestId('SearchInput'), 'iDoNotExist')
+
+      // Clear search field - onPress is tested in src/components/CircleButton.test.tsx
+      fireEvent.changeText(getByTestId('SearchInput'), '')
+
+      // Dapps displayed in the correct sections
+      const favoritesSection = getByTestId('DAppsExplorerScreenSearchFilter/FavoriteDappsSection')
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
+
+      // Names display correctly in the favorites section
+      expect(within(favoritesSection).queryByText(dappsList[0].name)).toBeFalsy()
+      expect(within(favoritesSection).getByText(dappsList[1].name)).toBeTruthy()
+
+      // Names display correctly in the all dapps section
+      expect(within(allDappsSection).getByText(dappsList[0].name)).toBeTruthy()
     })
   })
 
@@ -265,7 +362,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, getByText, queryByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -276,14 +373,14 @@ describe(DAppsExplorerScreenFilter, () => {
       expect(getByText(dappsCategories[1].name)).toBeTruthy()
 
       // Displays favorited dapp in Favorites section
-      const favoritesSection = getByTestId('DAppsExplorerScreenFilter/FavoriteDappsSection')
+      const favoritesSection = getByTestId('DAppsExplorerScreenSearchFilter/FavoriteDappsSection')
       expect(within(favoritesSection).getByText(dappsList[0].name)).toBeTruthy()
       expect(within(favoritesSection).getByText(dappsList[0].description)).toBeTruthy()
       expect(within(favoritesSection).queryByText(dappsList[1].name)).toBeFalsy()
       expect(within(favoritesSection).queryByText(dappsList[1].description)).toBeFalsy()
 
       // Displays other dapps in All section
-      const allDappsSection = getByTestId('DAppsExplorerScreenFilter/DappsList')
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
       expect(within(allDappsSection).getByText(dappsList[1].name)).toBeTruthy()
       expect(within(allDappsSection).getByText(dappsList[1].description)).toBeTruthy()
     })
@@ -300,7 +397,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -311,7 +408,7 @@ describe(DAppsExplorerScreenFilter, () => {
       expect(getByTestId('FavoriteDappsSection/NoResults')).toBeTruthy()
 
       // All Section should show only 'dapp 2'
-      const allDappsSection = getByTestId('DAppsExplorerScreenFilter/DappsList')
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
       // queryByText returns null if not found
       expect(within(allDappsSection).queryByText(dappsList[0].name)).toBeFalsy()
       expect(within(allDappsSection).queryByText(dappsList[0].description)).toBeFalsy()
@@ -331,7 +428,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -360,7 +457,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -396,7 +493,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -407,7 +504,7 @@ describe(DAppsExplorerScreenFilter, () => {
       fireEvent.press(getByText(dappsCategories[1].name))
 
       // Tap on remove filters from all section
-      fireEvent.press(getByTestId('DAppsExplorerScreenFilter/NoResults/RemoveFilter'))
+      fireEvent.press(getByTestId('DAppsExplorerScreenSearchFilter/NoResults/RemoveFilter'))
 
       // Assert correct analytics are fired
       expect(ValoraAnalytics.track).toHaveBeenCalledTimes(2)
@@ -433,7 +530,7 @@ describe(DAppsExplorerScreenFilter, () => {
       })
       const { getByTestId, getByText } = render(
         <Provider store={store}>
-          <DAppsExplorerScreenFilter />
+          <DAppsExplorerScreenSearchFilter />
         </Provider>
       )
 
@@ -443,7 +540,7 @@ describe(DAppsExplorerScreenFilter, () => {
       // Tap on category 2 filter
       fireEvent.press(getByText(dappsCategories[1].name))
 
-      // Tap on remove filters from all section
+      // Tap on remove filters from Favorite section
       fireEvent.press(getByTestId('FavoriteDappsSection/NoResults/RemoveFilter'))
 
       // Assert correct analytics are fired
@@ -456,6 +553,112 @@ describe(DAppsExplorerScreenFilter, () => {
         id: '2',
         remove: true,
       })
+    })
+  })
+
+  describe('searching and filtering', () => {
+    it('renders correctly when there are results in all and favorites sections', () => {
+      const store = createMockStore({
+        dapps: {
+          dappListApiUrl: 'http://url.com',
+          dappsList,
+          dappsCategories,
+          favoriteDappIds: ['dapp1'],
+          dappFavoritesEnabled: true,
+        },
+      })
+      const { getByTestId } = render(
+        <Provider store={store}>
+          <DAppsExplorerScreenSearchFilter />
+        </Provider>
+      )
+      // Search for 'tokens'
+      fireEvent.changeText(getByTestId('SearchInput'), 'tokens')
+
+      // Favorites section displays correctly
+      const favoriteDappsSection = getByTestId(
+        'DAppsExplorerScreenSearchFilter/FavoriteDappsSection'
+      )
+      expect(within(favoriteDappsSection).getByText(dappsList[0].name)).toBeTruthy()
+      expect(within(favoriteDappsSection).getByText(dappsList[0].description)).toBeTruthy()
+      expect(within(favoriteDappsSection).queryByText(dappsList[1].name)).toBeFalsy()
+      expect(within(favoriteDappsSection).queryByText(dappsList[1].description)).toBeFalsy()
+
+      // All section displays correctly
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
+      expect(within(allDappsSection).getByText(dappsList[1].name)).toBeTruthy()
+      expect(within(allDappsSection).getByText(dappsList[1].description)).toBeTruthy()
+    })
+
+    it('renders correctly when there are results in favorites and no results in all', () => {
+      const store = createMockStore({
+        dapps: {
+          dappListApiUrl: 'http://url.com',
+          dappsList,
+          dappsCategories,
+          favoriteDappIds: ['dapp1'],
+          dappFavoritesEnabled: true,
+        },
+      })
+      const { getByTestId, getByText } = render(
+        <Provider store={store}>
+          <DAppsExplorerScreenSearchFilter />
+        </Provider>
+      )
+
+      // Searching for tokens should return both dapps
+      fireEvent.changeText(getByTestId('SearchInput'), 'tokens')
+
+      // Filtering by category 1 should return only dapp 1
+      fireEvent.press(getByText(dappsCategories[0].name))
+
+      // Favorite Section should show only 'dapp 1'
+      const favoriteDappsSection = getByTestId(
+        'DAppsExplorerScreenSearchFilter/FavoriteDappsSection'
+      )
+      expect(within(favoriteDappsSection).getByText(dappsList[0].name)).toBeTruthy()
+      expect(within(favoriteDappsSection).getByText(dappsList[0].description)).toBeTruthy()
+      expect(within(favoriteDappsSection).queryByText(dappsList[1].name)).toBeFalsy()
+      expect(within(favoriteDappsSection).queryByText(dappsList[1].description)).toBeFalsy()
+
+      // All Section should show no results
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
+      expect(
+        within(allDappsSection).getByTestId('DAppsExplorerScreenSearchFilter/NoResults')
+      ).toBeTruthy()
+    })
+
+    it('renders correctly when there are no results in favorites and results in all', () => {
+      const store = createMockStore({
+        dapps: {
+          dappListApiUrl: 'http://url.com',
+          dappsList,
+          dappsCategories,
+          favoriteDappIds: ['dapp2'],
+          dappFavoritesEnabled: true,
+        },
+      })
+      const { getByTestId, getByText } = render(
+        <Provider store={store}>
+          <DAppsExplorerScreenSearchFilter />
+        </Provider>
+      )
+
+      // Searching for tokens should return both dapps
+      fireEvent.changeText(getByTestId('SearchInput'), 'tokens')
+
+      // Filtering by category 1 should return only dapp 1
+      fireEvent.press(getByText(dappsCategories[0].name))
+
+      // Favorite Section should show no results
+      expect(getByTestId('FavoriteDappsSection/NoResults')).toBeTruthy()
+
+      // All Section should show only 'dapp 1'
+      const allDappsSection = getByTestId('DAppsExplorerScreenSearchFilter/DappsList')
+      expect(within(allDappsSection).getByText(dappsList[0].name)).toBeTruthy()
+      expect(within(allDappsSection).getByText(dappsList[0].description)).toBeTruthy()
+      expect(within(allDappsSection).queryByText(dappsList[1].name)).toBeFalsy()
+      expect(within(allDappsSection).queryByText(dappsList[1].description)).toBeFalsy()
     })
   })
 })
