+++ src/dappsExplorer/useDappFavoritedToast.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { SectionList } from 'react-native'
 import ToastWithCTA from 'src/components/ToastWithCTA'
-import { DappV1, DappV2 } from 'src/dapps/types'
+import { Dapp } from 'src/dapps/types'
 
 const TOAST_DISMISS_TIMEOUT = 5000
 
@@ -10,7 +10,7 @@ const useDappFavoritedToast = (sectionListRef: React.RefObject<SectionList>) => 
   const { t } = useTranslation()
 
   // do not use favoritedDapp to show / hide the toast, as the content of the toast depends on the dapp name during transitions
-  const [favoritedDapp, setFavoritedDapp] = useState<DappV1 | DappV2 | null>(null)
+  const [favoritedDapp, setFavoritedDapp] = useState<Dapp | null>(null)
   const [showToast, setShowToast] = useState(false)
 
   useEffect(() => {
@@ -25,7 +25,7 @@ const useDappFavoritedToast = (sectionListRef: React.RefObject<SectionList>) => 
     sectionListRef.current?.getScrollResponder()?.scrollTo({ y: 0, animated: true })
   }
 
-  const onFavoriteDapp = (dapp: DappV1 | DappV2) => {
+  const onFavoriteDapp = (dapp: Dapp) => {
     setFavoritedDapp(dapp)
     setShowToast(true)
   }
