+++ src/statsig/constants.ts
@@ -73,13 +73,6 @@ export const ExperimentConfigs = {
       swappingNonNativeTokensEnabled: false,
     },
   },
-  [StatsigExperiments.DAPPS_FILTERS_AND_SEARCH]: {
-    experimentName: StatsigExperiments.DAPPS_FILTERS_AND_SEARCH,
-    defaultValues: {
-      dappsFilterEnabled: false,
-      dappsSearchEnabled: false,
-    },
-  },
   [StatsigExperiments.DAPP_RANKINGS]: {
     experimentName: StatsigExperiments.DAPP_RANKINGS,
     defaultValues: {
