+++ src/walletConnect/screens/useIsDappListed.test.tsx
@@ -9,23 +9,21 @@ const renderHookWithProvider = (dappUrl: string) => {
     dapps: {
       dappsList: [
         {
-          categoryId: 'finance-tools',
+          categories: ['finance-tools'],
           dappUrl: 'https://celotracker.com?address=0x047154ac4d7e01b1dc9ddeea9e8996b57895a747',
           description: 'Manage your Celo Portfolio from DeFi to NFTs',
           iconUrl:
             'https://raw.githubusercontent.com/valora-inc/dapp-list/main/assets/celotracker.png',
           id: 'celotracker',
-          isFeatured: false,
           name: 'Celo Tracker',
         },
         {
           name: 'Moola',
           id: '2',
-          categoryId: '2',
+          categories: ['2'],
           description: 'Lend and borrow tokens!',
           iconUrl: 'https://raw.githubusercontent.com/valora-inc/app-list/main/assets/moola.png',
           dappUrl: 'celo://wallet/moolaScreen',
-          isFeatured: false,
         },
       ],
     },
